+++ src/main/java/com/jcohy/scis/model/Student.java
@@ -23,7 +23,7 @@ public class Student  implements Serializable {
     private Integer id;
     //学号
     @Column(name = "num")
-    private Integer num;
+    private Long num;
     //密码
     @Column(name = "password")
     private String password;
@@ -55,11 +55,11 @@ public class Student  implements Serializable {
         this.id = id;
     }
 
-    public Integer getNum() {
+    public Long getNum() {
         return num;
     }
 
-    public void setNum(Integer num) {
+    public void setNum(Long num) {
         this.num = num;
     }
