+++ src/main/java/com/jcohy/scis/repository/StudentRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
  */
 public interface StudentRepository  extends JpaRepository<Student,Integer> {
 
-    Student findAdminByNum(Integer num);
+    Student findAdminByNum(Long num);
 
     Student findAdminByName(String name);
